bot/requirements.txt
google-generativeai>=0.8.0
google-genai>=0.2.2
supabase>=2.10.0
Pillow>=11.0.0
python-telegram-bot>=21.0.0
python-dotenv>=1.0.1
